app/requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.3
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
connectorx==0.4.5
gitdb==4.0.12
gitpython==3.1.46
idna==3.11
jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
matplotlib==3.10.7
narwhals==2.17.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
kaleido==1.1.0
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==6.33.5
pyarrow==23.0.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
qdrant-client==1.17.0
pyyaml==6.0.3
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wordcloud==1.9.4
